Move some tests to facilitate package renaming
PiperOrigin-RevId: 399923140
